style: move moment-format.ts to obsidian-markdown layer
Pure string transform with zero imports — belongs in the leaf parser
layer, not vault-operations (I/O layer).
